html/semantics/embedded-content/the-iframe-element/events.https.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  let log = '';
  function load(message) { log += 'LOAD: ' + message + '\n'; }
  function error(message) { log += 'ERROR: ' + message + '\n'; }
</script>
<body onload="load('body')">
  <iframe onload="load('invalid URL')" onerror="error('invalid URL')" src="https://:invalid-url"></iframe>
  <iframe onload="load('nonexistent')" onerror="error('nonexistent')" src="https://nonexistentsdfhrtesfkkjdfgoiuogd.com"></iframe>
  <iframe onload="load('MIX')" onerror="error('MIX')" src="http://{{host}}:{{ports[http][0]}}"></iframe>
  <iframe onload="load('cross-origin')" onerror="error('cross-origin')" src="https://{{domains[www1]}}:{{ports[https][0]}}"></iframe>
  <iframe onload="load('same-origin')" onerror="error('same-origin')" src="https://{{host}}:{{ports[https][0]}}"></iframe>
  <script>
    async_test(t => {
        setTimeout(
            t.step_func_done(() => assert_equals(log, '')),
            5000);
    }
              );
  </script>
</body>
